+++ src/test/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClientTest.java
@@ -1,10 +1,20 @@
 package com.getindata.connectors.http.internal.table.lookup;
 
+import java.net.URI;
 import java.net.http.HttpClient;
-import java.util.Properties;
+import java.net.http.HttpRequest;
+import java.util.*;
 
 import org.apache.flink.api.common.serialization.DeserializationSchema;
+import org.apache.flink.api.common.serialization.SerializationSchema;
+import org.apache.flink.configuration.Configuration;
+import org.apache.flink.formats.json.JsonFormatFactory;
+import org.apache.flink.table.api.DataTypes;
+import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.data.StringData;
+import org.apache.flink.table.factories.DynamicTableFactory;
+import org.apache.flink.table.types.DataType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,11 +22,14 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import static org.assertj.core.api.Assertions.assertThat;
 
+
 import com.getindata.connectors.http.internal.HeaderPreprocessor;
 import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
 import com.getindata.connectors.http.internal.table.lookup.querycreators.GenericGetQueryCreator;
+import com.getindata.connectors.http.internal.table.lookup.querycreators.GenericJsonQueryCreator;
 import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
 import static com.getindata.connectors.http.TestHelper.assertPropertyArray;
+import static com.getindata.connectors.http.internal.table.lookup.HttpLookupTableSourceFactory.row;
 
 @ExtendWith(MockitoExtension.class)
 public class JavaNetHttpPollingClientTest {
@@ -30,14 +43,18 @@ public class JavaNetHttpPollingClientTest {
     @Mock
     private LookupRow lookupRow;
 
+    @Mock
+    private DynamicTableFactory.Context dynamicTableFactoryContext;
     private HeaderPreprocessor headerPreprocessor;
 
     private HttpLookupConfig options;
 
+    private static final String BASE_URL = "http://localhost.com";
+
     @BeforeEach
     public void setUp() {
         this.headerPreprocessor = HttpHeaderUtils.createDefaultHeaderPreprocessor();
-        this.options = HttpLookupConfig.builder().build();
+        this.options = HttpLookupConfig.builder().url(BASE_URL).build();
     }
 
     @Test
@@ -57,6 +74,104 @@ public class JavaNetHttpPollingClientTest {
         assertThat(
             ((GetRequestFactory) client.getRequestFactory()).getHeadersAndValues())
             .isEmpty();
+    }
+
+    @Test
+    public void shouldBuildGetClientUri() {
+        // GIVEN
+        JavaNetHttpPollingClient client = new JavaNetHttpPollingClient(
+                httpClient,
+                decoder,
+                options,
+                new GetRequestFactory(
+                        new GenericGetQueryCreator(lookupRow),
+                        headerPreprocessor,
+                        options
+                )
+        );
+
+        DataType lookupPhysicalDataType = row(List.of(
+                        DataTypes.FIELD("id", DataTypes.STRING()),
+                        DataTypes.FIELD("uuid", DataTypes.STRING())
+                ));
+
+        RowData lookupRowData = GenericRowData.of(
+                StringData.fromString("1"),
+                StringData.fromString("2")
+        );
+
+        LookupRow lookupRow = new LookupRow()
+                .addLookupEntry(
+                        new RowDataSingleValueLookupSchemaEntry("id",
+                                RowData.createFieldGetter(
+                                        DataTypes.STRING().getLogicalType(),
+                                        0)))
+                .addLookupEntry(
+                        new RowDataSingleValueLookupSchemaEntry("uuid",
+                                RowData.createFieldGetter(
+                                        DataTypes.STRING().getLogicalType(),
+                                        1))
+                );
+        lookupRow.setLookupPhysicalRowDataType(lookupPhysicalDataType);
+
+        GenericGetQueryCreator queryCreator = new GenericGetQueryCreator(lookupRow);
+        LookupQueryInfo lookupQueryInfo = queryCreator.createLookupQuery(lookupRowData);
+
+        // WHEN
+        URI uri = ((GetRequestFactory) client.getRequestFactory()).constructGetUri(lookupQueryInfo);
+
+        // THEN
+        assertThat(uri.toString()).isEqualTo(BASE_URL + "?id=1&uuid=2");
+    }
+
+    @Test
+    public void shouldBuildBodyBasedClientUri() {
+        // GIVEN
+        DataType lookupPhysicalDataType = row(List.of(
+                DataTypes.FIELD("id", DataTypes.STRING()),
+                DataTypes.FIELD("uuid", DataTypes.STRING())
+        ));
+
+        LookupRow lookupRow = new LookupRow()
+                .addLookupEntry(
+                        new RowDataSingleValueLookupSchemaEntry("id",
+                                RowData.createFieldGetter(
+                                        DataTypes.STRING().getLogicalType(),
+                                        0)))
+                .addLookupEntry(
+                        new RowDataSingleValueLookupSchemaEntry("uuid",
+                                RowData.createFieldGetter(
+                                        DataTypes.STRING().getLogicalType(),
+                                        1))
+                );
+
+        lookupRow.setLookupPhysicalRowDataType(lookupPhysicalDataType);
+
+        SerializationSchema<RowData> jsonSerializer =
+                new JsonFormatFactory()
+                        .createEncodingFormat(dynamicTableFactoryContext, new Configuration())
+                        .createRuntimeEncoder(null, lookupPhysicalDataType);
+
+        BodyBasedRequestFactory requestFactory = new BodyBasedRequestFactory(
+                "POST",
+                new GenericJsonQueryCreator(jsonSerializer),
+                HttpHeaderUtils.createDefaultHeaderPreprocessor(),
+                HttpLookupConfig.builder()
+                        .url(BASE_URL)
+                        .build()
+        );
+
+        Map<String, String> urlBodyBasedQueryParameters = new LinkedHashMap<>();
+        urlBodyBasedQueryParameters.put("key1", "value1");
+        urlBodyBasedQueryParameters.put("key2", "value2");
+
+        LookupQueryInfo lookupQueryInfo = new LookupQueryInfo("{}", urlBodyBasedQueryParameters);
+
+        // WHEN
+        HttpRequest httpRequest = requestFactory.setUpRequestMethod(lookupQueryInfo).build();
+
+        // THEN
+        assertThat(httpRequest.uri().toString()).isEqualTo(BASE_URL + "?key1=value1&key2=value2");
     }
 
     @Test